delete

+++ src/controllers/clientes.controller.js
@@ -111,7 +111,7 @@ const clientesController = {
             };
 
             verificaCliente();
-            
+
         } catch (error) {
             console.log('Erro ao atualizar cliente: ' + error);
             return res.status(500).json({ message: 'Servidor não conseguiu realizar a operação.' })
@@ -233,12 +233,18 @@ const clientesController = {
             return res.json({ message: 'É necessário informar o id do cliente a ser deletado.' }).status(400);
         }
 
-        const queryDeletaCliente = 'DELETE FROM clientes WHERE id = ?';
+        const queryBuscaCliente = 'SELECT * FROM clientes WHERE id = ?';
 
         try {
+            const [rows] = await pool.query(queryBuscaCliente, ID);
+
+            if (rows.length === 0) {
+                return res.status(404).json({ message: 'Cliente não encontrado.' });
+            }
+
+            const queryDeletaCliente = 'DELETE FROM clientes WHERE id = ?';
             const [response] = await pool.query(queryDeletaCliente, ID);
             return res.status(200).json({ message: 'Cliente deletado com sucesso.' });
-
         } catch (error) {
             console.log('Erro ao tentar deletar o cliente: ' + error);
             return res.status(400).json({ message: 'Erro no contato com o servidor ou existem compras relacionadas ao cliente.' }).status(500);
